fix editing of multiple bodies' origins
Editing is at this point only possible in world coordinates, thus only
one side of a robot can be edited. We could include a list of of objects
along with an option to flip sides, but this would clutter the UI a lot.
However, the operator remembers the last numbers punched in, thus it's
easy to repeat the operator on other sides of the robot anyway.