implement stl export similar to obj export
A first try failed with exporting the stl objects as they were.
Apparently Blender does not simply export the mesh data from the
.data field of the objects, but takes into account the orientation.
Thus we need transform the mesh back to world space for export, which
is done similar to obj. The same axis settings are used as with the obj
for consistency.