fix error in assembling inertia for links
The change applied in #fec44126ff21532d66e69ec66cc90319bb5eb4f7 changed
the comparison to a 'inertial/name' instead of the obj.name property.
The problem with this was that inertial objects never have a name that
gets exported and thus never get one assigned, so that no inertial
objects were linked to a link.