Rename and update logging
The module logging.py masked the inbuilt logging library of python,
which is not used, but it's still bad style.
The commands startlog and endlog to identify the currently active
operator and pass it to logging were replaced by simply...passing the
active operator to the log function as the origin of the log call,
allowing for everything that was written only to terminal now to be
written to Blender's own log.