Refine logging for better feedback
To avoid unnecessary information in the Blender status line, the logging
as provided by phoboslog was adjusted. This improves the readability of
the returned information (e.g. for measure distance) by removing the
loglevel, as well as the log origin from this message.

A cleanup of unnecessary print commands was undertaken for
selectModelOperator, MeasureDistanceOperator and check_dict_alg in
validation.py.

The documentation of phoboslog.py and validation.py was completed.