Adjust export operator to be more comfortable
PEP8 formatting of some files.
Removed bug in selection.py (utils):
When no object was active, the poll-method of the set_mode operator
failed. Solution: use operator only when an object is active (if not it
should be object mode anyway).
Adjustments to export operator to:
a) show feedback on success as log
b) select all exported objects (making subsequent exports easier)