Split inertial creation into two operators
The inertial creation has been undertaken in step so far.
However, it is much more convenient to split the inertials into minor
and link inertial creation. This way it is possible to create link
inertials from selected minor inertials and experiment with different
inertia settings for each link.

With this commit the inertia operators have been moved to the
appropriate GUI section and split into minor and link inertial
creation.