Reimplement addHeightmap operator
The heightmap operator bug (link creation failed) is fixed by adjusting
the function call to involve the deriveLinkFromObject instead.
Additional parameters are added to support the subsurf modifier on
creation as well as the strength of the displacement.
Also, the object is shaded smooth.

Right now, the addHeightmap operator resides at the hardware tab, which
is not its final destination. As soon as the scene operators become
permanent, it should be moved there.

This commit also contains a commented GUI hack to switch to the modifier
context in the buttons area, however this does not work as intended. It
might be included in the future after it has been fixed.