Hotfix: introduce python installation script
The installation shell script has been around for a while. However, it
does not help for Windows users. This problem is resolved by introducing
a python installation file instead.

Also, the problems with the included YAML installation are resolved by
removing it completely from phobos. However, this requires the user to
do a proper installation of YAML. For this reason, the installation page
in the wiki was updated to provide all required information.

The Windows part of the installation file is not tested yet.

Fixes #126.