Add dynamic object information
The information of the active object is gathered by the getDictEntry
function of phobos.model.models.
Everything is sorted into subcategories and displayed in a neat layout
at the infopanel of Phobos.

This works as expected, however most of the link information is still
missing, as it is not provided by the model function.