Add dictionary unwrapping and GUI adjustments
Subcategories like joint/limits contain dictionaries which were not
printed properly. Now, these dictionaries are unwrapped and each setting
is shown individually in the form limits/upper etc.

Another change is the possibility to add operators to special properties
(like the joint type, visual geometry type etc.). This is already
included for the geometry type, joint type and motor type. The
extensible supportedProps allows further customisation.

The GUI is updated to the final look. As a consequence of the unwrapped
dictionaries the decision whether an item should be added to the left or
right column was improved. Now, the item is added to that column which
has less items at the moment, resulting in a clean and compact look of
the GUI.