Bugfix of failing definition parsing
The definitions of Phobos are parsed from the definitions folder.
However, this failed silently, as the defs.py contains only a relative
link to this folder, which was not found.

Move the definition folder to its proper place, so it will be copied
accordingly by the installation script.

Another issue was caused by the string formatting in the Addon bl_info.