Reconsider the configuration of additional modules
So far, we had used the yamlpath.conf to monitor the system path of
pythons site/dist modules. However, in the code we stripped the yaml
part anyway, so there is no reason to keep the file as "yaml-only".

Thus, we replaced the filename and the path which is saved in the file
to represent a system-specific (os.path.normpath) path where the
additional python packages are found.
This allows for future easy checking of requirements at the code
positions marked with "OPT" in __init__.py and install_phobos.py.

This should fix the Windows confusion mentioned in #132 and improve
future workflows. The gitignore is updated, too, to deal with the new
file naming.