Fix zero lenght axis problem with joints
When a joint axis is defined to be (0,0,0), the posebone is scaled down
to a point, which results in an exception when trying to set the
constraints.

This fix adds an error log, when this problem occurs, and ignores the
axis parameter in these cases.
That way, an URDF model can be debugged without throwing exceptions.