Fix entity, library, meshes and scenes initialisation
The plugin initialisations were using a python module (imp) which was
deprecated since version 3.4.
This might introduce errors, as Blender updates its API.

The required functions were replaced with a proper call to importlib.
(see https://docs.python.org/3.5/library/imp.html and
https://stackoverflow.com/questions/67631/how-to-import-a-module-given-the-full-path#67692)

The existing try-catch blocks were replaced with a unified check for
existing attributes.

The documentation was updated on the fly and a package docstring was
added.