Introduce submodel functions and definitions
The former assemblies are converted to a more general description as a
submodel. Submodel types are defined in the definitions/submodels.yml
file and are parsed to be provided to the user.

This allows future easy extendability of new submodel types like FMUs
etc.

When defined first, the respective objects are moved to a different
layer and the definitions from the yaml file are copied as custom
properties to the Blender group. See the docstring of the functions for
more details.