Fix AddSubmechanism Operator
There were numerous interacting problems in the operator, mostly caused
by properties being written to the objects defining the submechanism
even if the operator did not finish properly. The operator was
restructred to only make alterations to the scene if a proper mechanism
is defined, thus avoiding erroneous referencing etc.