Adjust link derivation to new inertia model structure
As we removed link inertial objects, we need to derive the inertia of a
link at export. This is done by the adjusted derive_inertia function.

This also cleans up the old code, as we do only need to call one
function for all different purposes.

Code cleanup on the fly.