Comment cryptic code part in deriveModelDictionary
This code excerpt is only used by submechanisms and will not work in its
current state with the restructured inertias.

As it is cryptic what this piece of code was supposed to do, it is
commented and added to the TODOs to figure out how to deal with it.

It appears it might not have worked properly before anyway.