Fix phoboslog issues when logging is disabled
There have been errors when deactivating the logging to terminal or
file. As logging information should be written to the UI (via
operator.report), there were errors when starting Blender with disabled
logging. These errors were caused by logging messages during module
registration, which could not be resolved to an operator call.

This was solved by introducing a utility function, which finds the
calling operator in the current stack frame and cancels if one of the
calling functions is in the FUNCTION_BLACKLIST. That way, logging in
anything else as an operator can be disabled when trying to log to the
Blender UI.

Furthermore, the file was style restructured and code was cleaned up.